requirements.txt
fastapi
uvicorn
openai
python-dotenv
chromadb
sentence-transformers
